send csv to backend, generate html from returned nodes
